feat: only copy symbolic links into output directory if asked to
Looking at GitHub Actions logs, it appears that the contents of
symbolic links are copied as the site is being built, which means that
a complete copy of the datasets is being created in each directory.
Since the lessons don't refer to these files directly (and can use
`..` if they need to), mccole is modified so that symlinks are only
copied on demand.

version: 0.2.1